Revert "dither: Precalculate dither noise as half floats"
This reverts commit 9270ef1caaccef23e82c7593eeb6e8e1898f2335.

+++ app/src/main/java/dev/kdrag0n/blurtest/GLUtils.kt
@@ -80,7 +80,7 @@ object GLUtils {
     }
 
     @SuppressLint("HalfFloat")
-    fun bitmapToRgb16fBuffer(bitmap: Bitmap, filter: (Float) -> Float = { it }): ShortBuffer {
+    fun bitmapToRgb16fBuffer(bitmap: Bitmap, filter: (Float) -> Double = { it.toDouble() }): ShortBuffer {
         return ByteBuffer.allocateDirect(bitmap.width * bitmap.height * 3 * 2).run {
             order(ByteOrder.nativeOrder())
             asShortBuffer().apply {
